test: qualify generic Gemma31 split-K routes
